Update docs.17.md (#1062)
Reworded for fluidity. I have long thought this page should be titled something else and at the moment I'm thinking Getting Started makes the most sense. "Introduction" tends to imply that a description/explanation is coming rather than a real-world example. That's of course my subjective assumption of it, but I would guess that my understanding is pretty common.